backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
gitpython==3.1.43
langchain==0.2.6
langchain-openai==0.1.13
langchain-community==0.2.6
chromadb==0.5.3
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
tiktoken==0.7.0
pydantic==2.7.4
httpx==0.27.0